fix: 詳細画面のproperties/property/idのパスを/properties/idへ修正
