+++ frontend/app/(main)/properties/[propertyId]/error.tsx
@@ -1,9 +1,10 @@
 "use client";
+
 import ErrorPage from "@/components/atoms/common/ErrorPage";
 import { MESSAGES } from "@/constants/messages";
 
-const PropertiesListError = () => {
+const PropertyDetailPageError = () => {
   return <ErrorPage errorMessage={MESSAGES.ERROR_UNEXPECTED} />;
 };
 
-export default PropertiesListError;
+export default PropertyDetailPageError;
