+++ frontend/app/(main)/properties/[propertyId]/loading.tsx
