+++ frontend/app/(main)/properties/property/[propertyId]/error.tsx
@@ -1,11 +1,10 @@
 "use client";
 
-const PropertyPageError = () => {
-  return (
-    <div className="h-[88vh] p-2 flex flex-col justify-center items-center text-center text-red-500 text-xl">
-      予期せぬエラーが発生しました。もう一度お試しください。
-    </div>
-  );
+import ErrorPage from "@/components/atoms/common/ErrorPage";
+import { MESSAGES } from "@/constants/messages";
+
+const PreviousPropertyDetalPageError = () => {
+  return <ErrorPage errorMessage={MESSAGES.ERROR_UNEXPECTED} />;
 };
 
-export default PropertyPageError;
+export default PreviousPropertyDetalPageError;
